# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/request/getOwnRequest": {
            "get": {
                "tags": [
                    "request"
                ],
                "summary": "get own request",
                "description": "if the user is trainer => get the requests sent to trainer\n     * if the user is a player  => get requests that this user sent",
                "operationId": "79d8d01d955abe5ef5cafb6c08d1fce3",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "type",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/request/send": {
            "post": {
                "tags": [
                    "request"
                ],
                "summary": "send new request",
                "operationId": "send new request",
                "requestBody": {
                    "description": "send new request",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "body",
                                    "trainer_id"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "lorem ipsum"
                                    },
                                    "trainer_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/request/show/{request_id}": {
            "get": {
                "tags": [
                    "request"
                ],
                "summary": "show request",
                "description": "show request",
                "operationId": "a9745c07912769cb655c5cadfc1ff0ab",
                "parameters": [
                    {
                        "name": "request_id",
                        "in": "path",
                        "description": "request_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/request/changeType/{request_id}": {
            "put": {
                "tags": [
                    "request"
                ],
                "summary": "edit request",
                "description": "edit request",
                "operationId": "20cf65b95c078e710964d736f07ed0ba",
                "parameters": [
                    {
                        "name": "request_id",
                        "in": "path",
                        "description": "request_id",
                        "required": true
                    },
                    {
                        "name": "request_type",
                        "in": "query",
                        "description": "request_type",
                        "required": true
                    },
                    {
                        "name": "trainer_id",
                        "in": "query",
                        "description": "trainer_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}